Game is fully functional, but minimally tested
* Refactored dice into new package
* Added util package
* Added scoring algorithm
* Added some debug/verbose mode output

This game actually plays farkle, and it turns out TerribleAi
isn't quite as terrible as I thought it might be... sometimes.
One time I saw it get 10500 in 22 turns which is a 477 average per
turn.  Of course, other times it was averaging around 35 turns to
go out, so there is still plenty of room for improvement.